+++ ninecms/__init__.py
@@ -1,7 +1,9 @@
 """ NineCMS package """
-__version__ = '0.5.3.1'
+__version__ = '0.5.4'
 __author__ = 'George Karakostas'
 __copyright__ = 'Copyright 2015, George Karakostas'
 __licence__ = 'BSD-3'
+
+from . import checks
 
 default_app_config = 'ninecms.apps.NineCMSConfig'